+++ app/src/main/java/ru/gidravpn/hydra/HydraApp.kt
@@ -1,0 +1,10 @@
+package ru.gidravpn.hydra
+
+import android.app.Application
+
+class HydraApp : Application() {
+    override fun onCreate() {
+        super.onCreate()
+        AppCtx.filesDir = filesDir
+    }
+}